#60 [REFACTOR] 키트 진단 커스텀 훅 분리
